travelgenie/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.5
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.128.8
google-auth==2.48.0
google-genai==1.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
jiter==0.13.0
openai==2.24.0
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==8.4.2
pytest-asyncio==1.2.0
python-dotenv==1.2.1
redis==7.0.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.49.3
tenacity==9.1.2
tomli==2.4.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
websockets==15.0.1
